requirements.txt
python-docx>=1.1.0
rerun-sdk>=0.12.0a4